# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Food Dispatch Backend API
  version: 1.0.0
  description: |
    Role-aware REST API for managing tenants, catalog entities, orders, and fulfillment
    flows for the Food Dispatch platform. All business-facing identifiers exposed to
    clients are sequential `ModelID` values (unsigned integers). Database `ID` values are
    never accepted from or returned to API consumers.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      additionalProperties: true
...
    Tenant:
      type: object
      properties:
        ID:
          type: integer
        model_id:
          type: integer
        Name:
          type: string
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
      required: [model_id, Name]
...
    User:
      type: object
      properties:
        ID:
          type: integer
        model_id:
          type: integer
        Email:
          type: string
        Name:
          type: string
        Role:
          type: string
          enum: [admin, tenant_owner, tenant_user]
        TenantID:
          type: integer
        Tenant:
          $ref: '#/components/schemas/Tenant'
        IsTenantOwner:
          type: boolean
      required: [model_id, Email, Role]
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
      required: [token, expires_at, user]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        tenant:
          type: object
          properties:
            name:
              type: string
          required: [name]
        user:
          type: object
          properties:
            email:
              type: string
              format: email
            name:
              type: string
            password:
              type: string
          required: [email, name, password]
      required: [tenant, user]
    RegisterTenantUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
      required: [email, name, password]
  responses:
    Unauthorized:
      description: Authentication required or invalid token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated user lacks the required role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register tenant owner
      description: Creates a tenant together with its owner user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Tenant owner created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: Authenticates a user and returns a JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/register-tenant-user:
    post:
      tags: [Authentication]
      summary: Invite tenant member
      description: Creates a secondary user within the caller's tenant. Tenant-owner tokens only.
      security:
        - BearerAuth: []
      x-roles: [tenant_owner]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTenantUserRequest'
      responses:
        '201':
          description: Tenant user created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: Invalidates the current session token.
      security:
        - BearerAuth: []
      x-roles: [tenant_owner, tenant_user]
      responses:
        '204':
          description: Logout successful.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
...